requirements.txt
streamlit==1.10.0
# Pillow==8.4.0
Pillow==9.2.0
# pandas==1.2.3
pandas==1.3.3